Updated the TDD to include steps for testing the Track Point Repository of the Train DB and template to describe the Adjacent Point Repo
Updated the TDD to include steps for testing the Track Point Repository
of the Train DB and template to describe the Adjacent Point Repo. Also,
improvements were made on the description of how Junit will be launched
to run tests and a more detailed explanation on how the train position
estimation algorithm is tested.