Fixed bugs detected from second round of integration testing
Added Debug code to write to file all of the bytes that are received
from the MDU.
Changed the formation of trainId generation to no longer have leading
zeroes as they cause a problem for the restsful service lookup.
Fixed a bug where the time sync response sent to the MDU not calculated
correctly (nanoseconds instead of seconds were attempted to be sent)
Fixed bug where train identifiers that do not have a train position yet
can be advertized in the known train identifiers list of the Train
Navigation Service.
Fixed bug where the train position service of the train nav service
restful service client expected integer train identifiers instead of
String ones.
Added additional error message to clarify any failure to report position
because a position is not available when looking up a train by its train
identifier.